requirements.txt
aioeasywebdav==2.4.0
aiohttp==3.6.2
alembic==1.4.1
appdirs==1.4.3
async-timeout==3.0.1
attrs==19.3.0
bcrypt==3.1.7
biopython==1.76
boto3==1.12.20
botocore==1.15.20
cachetools==3.1.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
ConfigArgParse==1.0
cryptography==2.8
datrie==0.8
decorator==4.4.2
docutils==0.15.2
dominate==2.5.1
dropbox==9.4.0
filechunkio==1.8
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
ftputil==3.4
gitdb==4.0.2
GitPython==3.1.0
google-api-core==1.16.0
google-auth==1.11.2
google-cloud-core==1.3.0
google-cloud-storage==1.26.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
idna==2.9
importlib-metadata==1.5.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.9.5
jsonschema==3.2.0
jupyter-core==4.6.3
Mako==1.1.2
MarkupSafe==1.1.1
multidict==4.7.5
nbformat==5.0.4
networkx==2.4
numpy==1.18.1
pandas==1.0.2
paramiko==2.7.1
prettytable==0.7.2
protobuf==3.11.4
psutil==5.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.20
Pygments==2.6.1
pygraphviz==1.3
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pyrsistent==0.15.7
pysftp==0.2.9
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
python-irodsclient==0.8.2
pytz==2019.3
PyYAML==5.3
ratelimiter==1.2.0
redis==3.4.1
requests==2.23.0
rq==1.3.0
rsa==4.0
s3transfer==0.3.3
six==1.14.0
smmap==3.0.1
snakemake==5.10.0
SQLAlchemy==1.3.15
toposort==1.5
traitlets==4.3.3
urllib3==1.25.7
visitor==0.1.3
Werkzeug==1.0.0
wrapt==1.12.1
WTForms==2.2.1
xmlrunner==1.7.7
yarl==1.3.0
zipp==3.1.0
